+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 # alpha2 / dev-master
 
-- Added `--pretty` option to `session:export:view` command to output formatted XML
-- Ask confirmation before overwriting file in `session:export:view`
+## Features
+
+- New command: `file:import`: - import files into the repository.
+
+## Improvements
+
+- `session:export:view`: Added `--pretty` option to `session:export:view` command to output formatted XML.
+- `session:export:view`: Ask confirmation before overwriting file.